Implement trigger require/reject-approvals
A pipeline can have triggers from more than one connection, thus it is
necessary for the github driver connection triggers to support a trigger
level require and reject capability on reviews. This change implements
that capability.

There was a small logic flaw in model.py that was blocking a trigger
that only has a reject-approval filter and the change itself does not
have any approvals.

Change-Id: Ia2471381153892e1083e09534d643de3e2bfb76f