Abort jobs if JobDir disk usage exceeds an optional maximum
This adds an optional setting to the [executor] section that allows
admins to set a limit on how much local disk space a job can use before
it is aborted.  A new watchdog thread is spawned, similar to the existing
timeout watchdog, to monitor the running job and abort when the
limit is met.

This also updates the zuul.conf sample with the missing [exector] section.

Story: 2000879

Change-Id: I23b0f85673cb6e1ab71b5912507799e7cea178fb