Expose ref to the executor and url formatter
For github (and probably other providers) we really only have the option
of returning 1 url for the entire buildset, as opposed to 1 per build.
To make log uploading within that easier we really need a way to
globally identify all the different builds that belong to 1 change.

It turns out we already have this in the form of a zuul ref. Currently
this information is really only used in the cloning and merging process
though. Expose it as a parameter to the builder and the url formatter so
that we can reference it from a reporter.

Change-Id: I1cab8af5c9d7f6875591fbe4ac4e184b90f6ca12
Signed-off-by: Jamie Lennox <jamielennox@gmail.com>