make Job and ZuulRole hashable
Python3 removes the default __hash__ method on object if a cusotm __eq__
is defined. This is to encourage authors to make things that are equal
more likely to have the same hash.

For now we keep python2 behavior (which is to use id(self). Howeve we
should probably make it a real hash of the object that matches __eq__'s
comparisons.

Change-Id: I01c9569bd2d196bb4ffa38276695fefe3c72c957