Update patch set 15
Patch Set 15:

James E. Blair wrote:

> Should we update gerritdriver to support this as well?
>
> Do we really want the config option, or should we just always have
> the change filter?

Well, gerrit doesn't seem to use the status_url per se, it uses the pipeline start-message to format the comment it makes if so configured.

After some debate on IRC, it boils down to this: We have a need to configure a URL that will be sent in the body of a Github status. It seems that this URL is really Zuul site specific, not something that would change per pipeline or per connection. That said, you explicitly removed the status_url from zuul config, to break a connection between reporter and scheduler. Thus, we're unsure where the appropriate place would be to re-introduce the configuration.

Once where to put the configuration of the URL itself is sorted out, then I think it's much easier to just simply make that URL configuration always use the url with the change tacked on.

Does any of that make sense?

Patch-set: 15
Reviewer: Gerrit User 6589 <6589@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0