Update patch set 15
Patch Set 15:

Yep, gotcha.  The gerrit/difference here is that the status url for github is not necessarily part of the message; it's structured metadata.

I think it may make the most sense to set status_url in zuul.conf.  Always include the change filter (no separate option for that).  It's going to need the tenant as well, at some point.  Automatically include that in the github metadata.  Also make it available as an format string variable for use in the pipeline start-message.

How's that?

Patch-set: 15