Don't report start on unmanaged projects
When using a check pipeline with enabled start reporting zuul reports
start on all projects of the gerrit instance regardless if they are
part of the pipeline or not. After that zuul notices that there are no
jobs to do.

The fix is to delay reporting until zuul knows if there are jobs to
run. This also implies changes to reporting of syntax errors such that
zuul now just reports with "syntax error" instead "starting jobs" and
then almost immediately "syntax error".

Change-Id: Idd754f3e42157d50012f49ef89cfb05f566efd7e