Convert some leftover changeish mentions to ref
Noticed there were a few vestigal references to Changeish that didn't get
converted, including the ChangeishFilter. Convert them to Ref.

In model, there were two places where a method started with changeish =
self.change. Those were removed and self.change was used in the method
because ref = self.change led to the method having two important
variables "ref" and "ret" which seems like a recipe for confusion. Of
course, that led to one line being too long for pep8, so it was
rewritten to use getattr instead of hasattr + and.

Change-Id: I420b8ec7c7eb32574dfe2de467fc2bbd0f96e9f1