Update patch set 3
Patch Set 3: Code-Review-1

The console server is intended to continue to run safely across jobs.  We can't be certain that we can kill it in all cases at the end of the job.  If we tried to kill it at the start of a job, we might do so after a user connected to it (and besides, we want to start as early as possible).  Instead, it's easier to just make it safe to always be running.

Patch-set: 3
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1