Update patch set 2
Patch Set 2: Code-Review+1

(5 comments)

This is pretty cool. However, a Dockerfile in the root of the tree is a pretty large implication of "this is the way to Docker zuul". I wonder if we should consider moving that file to tools so we don't endorse this too hard when only one of our users actually uses it that way. Or maybe not.. I'd love to hear peoples' thoughts on that.

Patch-set: 2
Label: Code-Review=+1