Update patch set 4
Patch Set 4: Code-Review+2

Joshua, yes, that's still a possibility.  And Jamie has a change staged to add that back.  But this change is simple and is forward-compatible with that, so I think we should land it.

Patch-set: 4
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2