Update patch set 1
Patch Set 1:

This is related, but not a -1: for the case of generating the log urls, I'd like to add a provision to the executor to read data back from completed jobs.  We can have the log-copying role write the URL for the logs into a JSON file in the job working directory, and then have zuul load that json file and make data from it available to the reporters.  We can also define a standard variable name for the build log path, so that we don't have to use success-url or failure-url in the job definition.

That's still awkward for the case of a *buildset* url rather than a build url, so we may still need the facility you are developing here (or we could just have zuul use the first build log url as the buildset log url).

Patch-set: 1