Update patch set 2
Patch Set 2:

Note, this will cause the start message to be delayed until the item ends up inside of the active window.  Part of the reason we added the start message was so that users got feedback when the system was backlogged.  This reduces the effectiveness of that -- it changes the semantics from "this change is in the queue" to "jobs are running on this change".

Patch-set: 2
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0