Update patch set 1
Patch Set 1: Code-Review-1

The cloner is not going to exist in v3 in the same way as v2.  The executor now has the reponsibility for setting up the git repos for a job.  Most of the branch logic will move to there.  If we retain the cloner as a separate command, it will be only so that users have a way of locally reproducing the build state for a job after it has already run.  What we are likely to do with it is simply have it repeat the actions already performed by the merger, after recording them to some kind of data structure.

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1