Update patch set 2
Patch Set 2: Code-Review-1

Hrm. This opens up a can of worms in my head. The SQL reporter is a plugin, so the configure connections part of it zuul shouldn't know about it specifically. (the SMTP reporter would be the same- and I think a FedMsg reporter would also be the same)

I'm wondering if we shouldn't extend the ConnectionInterface to take the name of the service (scheduler, merger, executor) and define an enum somewhere of services so that connections can filter which services they're needed by.

I think we should loop jeblair in on this one.

Patch-set: 2
Label: Code-Review=-1