Update patch set 1
Patch Set 1:

Agreed this should not be needed moving forward, but in the meantime there are jobs that use ZUUL_ variables and it would make the migration easier if we could keep the legacy environment.

For example, DLRN and rpmbuild jobs are using ZUUL_ variables to dynamically clone distgits associated with projects.

Patch-set: 1
Label: Verified=0