Update patch set 3
Patch Set 3: Code-Review-1

Neat! I think the idea of this is great!!

A few suggestions and/or things I think we need to think through together:

We've already got an HTTP interface (where status.json is served) Instead of adding the pecan, since this is so far a fairly simple interface, can we just add the pecan controllers as methods to the existing http interface? Or - if there is a real win we're getting from pecan here so that we do need it for this, then perhaps we need to update the entire current interface. (I mostly think having two different ways in zuul of serving http content seems like not a good plan) We should also have a quick discussion as to whether, if zuul is to grow an attachment to a framework like pecan, if pecan is the 'right' choice. Zuul is not beholden to OpenStack patterns - if it was me a vacuum today I'd likely choose flask. BUT - I'm not actually suggesting one over the other ... I mostly want to make sure as we add such a dependency that we talk about the pros and cons.

Relatedly, and this is based largely on my experience writing REST API client things for OpenStack - I think we should add a /capabilities endpoint that can return information indicating whether or not the zuul installation in question supports the dashboard API (since it will not if it doesn't have a SQL reporter) That capabilities endpoint could be used in the status page to know whether or not to show a link to the dashboard page - or by other people for other purposes. But having a part of an HTTP API that is only there sometimes is one of my least favorite experiences in OpenStack.

I believe we'll also want to consider how this works with v3 multi-tenancy.

Finally - I started a tad concerned with the angular bit - largely because other angular things I've seen come across like a ton of black magic. HOWEVER - I'd like to say that *this* angular is actually extremely clear and it makes sense to me why it's desirable. I'm potentially on board with it - however if we do decide that this is the way this should be done I think we should investigate aligning the status page - and also probably our log-streaming interface so that we have a clear answer for how such features should exist.

Patch-set: 3
Label: Code-Review=-1