Update patch set 3
Patch Set 3:

Thank you for the comments! Note that this is a proof of concept that happens to be enough to roughly replicate the Jenkins web interface. If this is deemed worthy for the Zuul project, then I guess it needs to be specified and better integrated... In the meantime:

I made it a separate process so that it can be deployed independently from the scheduler. I also have concern about its performance with a production workload where the whole table needs to be crunched to get the job lists.

Whenever it is served by the zuul.webapp or standalone, the HTTP API needs to be carefully designed indeed :) For example, it would be convenient to also add a "retrigger" button in the job list, in particular for post/tag jobs. That would give us:

* GET /capabilities : return available endpoints
* GET /status
* GET /job : return jobs list
* GET /job/$job-name : return job runs
* PUT /build/$buildset-id : re-enqueue a job

Then we may want results pagination, tenant's scope and surely a CLI client... How about using swagger then?

Lastly, regarding the libraries choice, I picked pecan/angular simply because I'm more comfortable with them, but I've no strong opinions and this could easily be rewritten.

Patch-set: 3
Label: Verified=0