Update patch set 1
Patch Set 1: Code-Review-1

We're generally not maintaining backwards compatability for running jobs, instead we expect to transform current jobs into v3 jobs (either by hand or with the help of scripts).  The shell vars are a grey area, since it's not as easy for us to mechanically translate job content (shell scripts) as it is job metadata (jjb -> zuul).

If we keep these variables in v3, it will be *very* difficult to remove them, and, moreover, they'll still be present in every v3 job, where we are trying to make the environment for running ansible as clean as possible.  I'd rather we translate the zuul ansible variables into shell variables in the legacy jobs that need them, so only those jobs are affected.

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1