Update patch set 1
Patch Set 1:

Ok - what about this as a solution:

Make an ansible module called "zuul_legacy_envvars" that takes as a parameter the zuul vars dict that are in the ansible vars and returns a dict with ZUUL_ envvars.

Then as part of the mechnical transform of old JJB jobs to new ansible playbooks, if a playbook has shell tasks (I expect all of them will) we add a task at the top which is:

  zuul_legacy_envvars:
    zuul_vars: "{{ host_vars[ansible_host].zuul }}"
  register: zuul_env

And then write the shell sections as:

  shell:
    echo "Legacy ZUUL_REF is $ZUUL_REF"
  environment: "{{ zuul_env }}"

That'll allow for a backwards compat mechnical transform -but will also make it clear that the step of getting ZUUL_ env vars is a legacy thing - and should be easy to grep for and remove as job content becomes more directly able to consume the new zuul vars.

Bonus points in the jjb->ansible transform could look for $ZUUL_ in the shell content and only add the task/environment line as needed - although that could be error prone if there are external shell script files that are being called that themselves use $ZUUL_ vars - so the safest thing would be just to add it to all shell calls in the transform then remove them over time.

Patch-set: 1