Strip unneeded cmd, changed, stdout and stderr from results
Logging gets unreadable because each command run winds up reporting back
a poorly formatted version of the entire script, which is often -x'd
anyway, as well as the full stdout, which we just streamed into the log.

Strip these so that the result dict is clear and contains useful info.

A followup should likely take start, end, delta and rc and make a nicely
formatted single message.

Change-Id: I57b6f5a12474e75a7f30a4e95e93a007965a2f13