Rename 'repos' job attribute to 'required-projects'
The name 'repos' is not clear, especially since we have removed
most other occurences of that word.  It's not obvious that the
values should be names of Zuul projects.  We might simply use
'projects' instead, but there is a concern that users may see
that as saying this job should *run on* these projects (ie, as
a replacement for the project-pipeline configuration).  Hopefully
'required-projects' indicates that the job utilizes the projects
in the list.

Change-Id: I43af7a7f688f1368514427a2892a03d1cf46fe72