Enable test_multi_branch cloner/executor test
This changes the method of evaluation to verify that commits are
present rather than checking for ZUUL_COMMIT -- that variable is
no longer correct as, in the case of a merge commit, it represents
the SHA which was generated on the merger, which will not be the
same SHA generated on the executor.

The *content*, however, will be the same.

Change-Id: I4bed780c75ef9e586b3cfce2a88322dcaf9df1d5