Make finger port configurable
In some setups (e.g. unprivileged container) it is not allowed to bind
to privileged ports. In order to still being able to use the log
streamer the finger port must be configurable. In case of a docker
container the port could be mapped to the default port outside then.

Change-Id: I996b4740876a71d99cc4d22dda6beda5b9b2f028