Remove test_upgrade
This test used the cloner to check out a set of git repos twice.

This does not map to the executor since it will only check out
git repos once.  To perform an upgrade test in the future, jobs
will need to make their own copies (or changes) to the git repos
provided by the executor.  Zuul v3 makes this easier since all
of the git repos it provides have all of the proposed future
branch states, so within an upgrade job, it is as simple as running
'git checkout oldbranch'.

Change-Id: I8c5ba0acaa5efaf392fec94eaecf68b263e945ba