Add github reporter status-url config option
Make it possible to set the github status link to something other than the
zuul server status_url. This allows the github status to link to the log output
for the jobs. If set, this url will be run through the same formatter
as the url_pattern used for comment links. If unset, it will fall back to the
previous behaviour of using the zuul server status_url.

Change-Id: I5f2ff35bb38426ddcd7f65798b4f114256a54847