Update patch set 1
Patch Set 1: Code-Review+1

@SpamapS - the follow up change does exercise this code.

The github reporter reports when all the jobs are done (the buildset) and so uses formatUrlPattern without having a specific build or job that it can reference. That is the driving reason for this change. 

It's likely that this patch and the follow up should be merged, but they weren't in the patches in our tree so they got proposed here in the same way.

Patch-set: 1
Reviewer: Gerrit User 7191 <7191@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1