Update patch set 1
Patch Set 1:

I understand the use case and why something like zuul_changes is required to handle it -- essentially, you can have a job with many repos with potential changes, but unlike a devstack job where you simply install all of them from git, you need to know which ones to build because building is expensive.

Unfortunately, zuul_changes can't really be ported over directly because so much of how we reference projects is changing (to accomodate multiple sources and multiple connections -- for example, a build that depends on both a gerrit and a github change is possible).  However, we can create a new piece of data that expresses what we need, and then, to facilitate compatibility, implement an ansible module as Monty suggests to translate that to zuul_changes if that's needed for a transition.

What I think we should do in zuul is create a new detailed data structure in ansible variables that fully describes the changes involved in a build.  Something like:

vars['zuul']['changes'] = [
  {'project': {'canonical_name': 'git.openstack.org/openstack/nova',
               'name': 'openstack/nova',
               'connection':
                    {'name': 'gerrit', 'canonical_hostname': 'git.openstack.org'},
  'branch': 'master',
  'number': '1234',
  'patchset', '1'},
  ...
]

That should allow us to supply all the relevant information to a job.  Using that information to perform the build steps you describe should be straightforward, and it can still be converted into the ZUUL_CHANGES form (while losing a lot of information) if needed.

Patch-set: 1