Update patch set 3
Patch Set 3: Code-Review-1

Okay, I'm +1 on the change in general, but I still think the "buildset zuul ref" as a concept may be going away.

However, you're not really using it as a *ref* here, you're just using it as a uuid.  I'd rather we go ahead and declare that buildsets will have uuids (we can drop the initial "Z") and start referring to them as just that, rather than as zuul refs.

As a transition, for now, let's set the buildset zuul ref to "Z<buildset uuid>" so we end up with the same behavior and can make this change minimal.

Patch-set: 3
Label: Code-Review=-1