Remove ansible files at startup before copy
When starting executor a second time, a traceback was happening because
the destination for the ansible files already existed. shutil.copytree
requires that the destination path does not exist. Since what we seem to
want is a fresh copy of the ansible files at start up, and no left over
cruft, removing the tree seems correct.

Change-Id: I92418f0e6233558f8aaee796a679d638f7c4c4c7