Update patch set 4
Patch Set 4: Code-Review-1

(1 comment)

The one second pause makes this pretty agressive, do we want to extend that delay?  Performing all of these stat calls may be fairly resource intensive, especially when you consider that an executor may have 100 of these threads running at the same time.  It could very easily end up doing nothing other than this.

In fact, this concern is what led me to suggest in the story[1] that rather than having a thread for every job, we just have a single thread on the executor, and have it iterate through all of the jobs.  That way, there is never more than one thread performing stat calls.  We're likely to have a significant delay between checks for a given job, but we will regain quite a bit of performance.  Is there a reason to go with this approach rather than the one I suggested?

[1] https://storyboard.openstack.org/#!/story/2000879

Patch-set: 4