Update log to include expected data on task results
As a followup to the previous patch, add task results back to the
output. This can't be done by delegating to the base class as the base
class uses the display methods and we can't update their format.

Change the format to always be timestamp first to be consistent with the
existing 2.5 logging formats.

Output now looks like:

  2017-06-07 14:00:53.552257 | PLAY [foo/test : all]
  2017-06-07 14:00:53.567250 | TASK [shell: echo foo]
  2017-06-07 19:00:54.942999 | node.example.com | foo
  2017-06-07 14:00:55.146433 | node.example.com | ok: Runtime:
    0:00:00.003718 Start: 2017-06-07 19:00:54.940639 End: 2017-06-07
    19:00:54.944357
  2017-06-07 14:00:55.151159 | TASK [Output important information]
  2017-06-07 19:00:55.971276 | node.example.com | bar
  2017-06-07 14:00:56.231973 | node.example.com | ok: Runtime:
    0:00:00.003748 Start: 2017-06-07 19:00:55.968980 End: 2017-06-07
    19:00:55.972728
  2017-06-07 14:01:52.283504 | TASK [Get status of pydistutils.cfg file
    path={{ '~' | expanduser }}/.pydistutils.cfg]
  2017-06-07 14:01:53.517179 | node.example.com | ok: Results: =>
    {"changed": false, "stat": {"exists": false}}
  2017-06-07 14:01:56.238066 | TASK [shell: exit 1]
  2017-06-07 14:01:57.324672 | node.example.com | ERROR: Results:
    => {"delta": "0:00:00.003457", "end": "2017-06-07 19:00:57.007769",
    "failed": true, "rc": 1, "start": "2017-06-07 19:00:57.004312"}

Wrapped indentation is for the commit message - the lines are each
single lines. The playbook was invoked as "foo/test.yaml" - so the first
entry shows the beginning of a play from that playbook. The first task
is just the bare shell command "echo foo". The second task has a name.
The third shows a non-shell task and the final task is an error.

It should be noted that in this case the timezones do not match on the
executor and the remote node. (it's my laptop and a cloud server)

Change-Id: I0159b08565dd3b08efd1cdec86ad5aa26bc2988e