Allow per-repo selection of configuration classes to load
So that multiple Zuul installations can share portions of their
configuration, allow the administrator to indicate which
configuration objects should be loaded from which repositories.

This also facilitates third-party CI, and is important for
any interaction with repos for which a given Zuul installation is
not fully responsible.

In particular, this allows an administrator to use the jobs, but
not the project-pipeline definitions from a given repo.  Or even
to use the content of a repo without reading any of the zuul
configuration therein.

Change-Id: I8a07e298c8cf4dd7cbf6f5b7fc38990f7d740af4