Fix git ssh env in merger
A recent change switched from using .ssh_wrapper files to using
a custom environment to set the ssh command in the merger.  However,
it was incomplete.

Ensure that all instances of git.Repo objects have the necessary
environment variable set.

Change-Id: Ie69c736ded1628e5ccf7afd4535d31c6c9b20215
NB: this feature (enabled in the earlier change) requires Git 2.3+.