Extract get_playhosts listing in zuul_stream to a method
We have a TON of jobs that only run on one host. For those, prepending
the hostname into the log line is noise. Only add it if it would be
information for the job.

We're not ready to pull the trigger on doing that yet, but this refactor
makes it easy to do so when we are.

Change-Id: I8d6073d35a9d885341f95a8b7b6d9469ac280c3f