Add build.started state flag
Add a 'started' boolean rather than relying on the existence of url
since url is computed.

As part of this, stop reporting manager as distinct from worker. They're
ultimately the same thing and the word 'manager' in the codebase refers
to the Pipeline manager, so it's not helpful.

Change-Id: I96898e6f951766da281bbd8dfa3086a0648baffa
Story: 2000782
Task: 3306