Update patch set 1
Patch Set 1: Code-Review-1

There are numerous events that we don't specifically handle. https://developer.github.com/webhooks/#events If we want to be "nice" about it and not return a 5xx to github, we should add some catch-all code so that the things we do not handle get quietly discarded and a success is returned back to github.

Patch-set: 1
Label: Code-Review=-1