Update patch set 1
Patch Set 1: Code-Review-1

This raises some questions for me; the -1 is mostly because I want to make sure we have answers before we merge this.

We run ansible inside of its own process group so that we can effectively kill it in the case of a timeout.  It looks like this change doesn't relate to timeouts, but rather to a process that gets spawned in the background.

Since this change only addresses bubblewrap, it only affects untrusted playbooks for now (we have discussed perhaps using bubblewrap for trusted playbooks, but don't at the moment).  If it seems reasonable to say that even trusted playbooks should not expect to be able to start daemons, then I wonder if we should look at doing the same for trusted playbooks?

Could we achieve the same result for both kinds of playbooks by performing a process group kill at the end of the playbook in all cases (ie, not just in the case of a timeout)?  If so, would we not need this bubblewrap option?

Patch-set: 1
Label: Code-Review=-1