Move streaming url formatting to model
We're currently formatting the in-progress status URL in the executor
client. However, once we add websockets streaming, we're going to have 2
potential URLs to report to a user. Rather than make the executor need
to know about all of that, stop formatting the url in the executor
client and move it to the model. Just report the information that's
needed from the executor back to the client so that the client can deal
with presentation.

Change-Id: I43fe5d28b56848c62c4fcfcf957755f83fdc1c78