Update patch set 1
Patch Set 1:

So the problem is that on the Github side there is an ugly error shown if a user happens to look at their repo's web hooks configuration (if they're using a direct web hook config instead of a proper Github integrated app). There's nothing the user can do to prevent the appearance of errors, which is poor UX. I think it would be better for zuul to accept anything Github throws at it, and just quietly discard the things it will not act upon.

Patch-set: 1