Update patch set 8
Patch Set 8: Code-Review-1

I actually think we should continue to format the streaming url inside the executor, but I agree we should not do it in two places.  When you consider what this will look like on completion, we're going to have:

  User -> multiplexer -> log_streamer

The "finger" url of the log_streamer will eventually only need to be accessible to the multiplexer.  I think we're eventually going to hide that from the user.  Since the executor absolutely knows what the finger URL should be (hostname_or_ip + port + job id -- and yeah in containers hostname_or_ip may actually need to be a config option, but that's fine), it's actually best placed to format that url.  In other words, since we don't expect *this* connection to be going through any gateways, load balancers, whatever, let's just let the executor format the "url" for the job as it has been, and treat it as an internal url.

Now, once we have the multiplexer, that's where ui concerns come in to play.  We should be able to specify both a finger url and a websocket url for a job.  I think *that's* what should be configured in the jobs and formatted in the model.  It would probably look like:

  web-streaming-url: ws://zuul.example.com/{build.uuid}
  raw-streaming-url: finger://zuul.example.com/{build.uuid}

When the multiplexer gets one of those requests, it looks up the build, finds the internal streaming url, and proxies the connection.

But until we have the multiplexer, let's stick the internal url on the status page so we can continue to debug.

How does that sound?

Patch-set: 8
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1