Update patch set 8
Patch Set 8:

I agree with most of what you say there and think it's spot-on, but I'm still not sure we need to do any actual url string formatting in the executor since the only consumer of a url would be the end user.

To generate a finger URL, one needs a triplet of hostname|ip, port and resource.

The multiplexor and the websocket also need that same triplet to connect to the internal service.

The executor already returns the hostname|ip of the node to the scheduler - and the scheduler already has the uuid of the build. If the executor returns the port as well, then we have the information all the consumers need.

We only need the url when we're going to put it in the status.json body

It seems to me that it's important that the executor return the triplet of data to be stored in the model class. Then when a finger location needs to be displayed in the status, it can be rendered into a URL, and when it needs to be used for internal communication, it's available to be used directly.

Patch-set: 8
Reviewer: Gerrit User 2 <2@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0