Update patch set 1
Patch Set 1: Code-Review-1

So I don't have a concern with handling ping, but agree with jlk. There are a lot of webhooks that zuul doesn't handle and that doesn't necessarily mean that it was an error. 

I think I would just prefer to change the default here so that an unhandled event returned an empty 200. It's not an error to get those events and if we choose to ignore them that's no reason to not respond that we successfully handled them.

Patch-set: 1
Label: Code-Review=-1