Update patch set 2
Patch Set 2: Code-Review-1

This is fantastic and I think will be very useful.

There is an important consideration though, which is that in zuul.yaml files, order is important - at least as related to the definitions of multiple variants of a job.

I think we should sort the directory listing, so that ordering is deterministic.

We can eventually write some docs explaining that a user may want to keep all variants of a given job in the same file. However, sometimes that may not be desirable. For instance, one might want to define jobs in one file and stable/ branch variants in another. In that case, a user could prepend a number or something similar to their file to ensure the main entries got loaded first.

In any case, I think adding a sort should be pretty easy. Otherwise this patch looks great!

Patch-set: 2
Label: Code-Review=-1