Update patch set 3
Patch Set 3: Code-Review+2

(1 comment)

Tristan Cacqueray wrote:

> Glad configuration split is still a thing :-)
>
> @James, I added a first test, and if it's ok with you, I think the
> best would be to modify some of the existing tests fixture layout.
> Also, when adding this first test, I had to modify the
> executor/server that seems to be handling the cat job instead of the
> merger/server.

I'm happy with this test, but if you want to modify a few others to use the split, that's fine too.

And yeah, there is currently duplication of merger jobs in executor/server and merger/server.  Unfortunately, they both need to be the same, as they both may be used in production, though we only exercise merge functions in executor/server in tests.  We should perhaps abstract those into a shared helper class to avoid the duplication.

> @Monty, Indeed I figured this isn't going to play well with job
> variants. So I added the sort you suggested but it seems like the
> configloader should be more rigorous when loading split
> configuration...

We *could* be more rigorous (e.g., disallow variants from different files in a split config), however the scenario Monty described may be useful, so I think it's worth trying the sorted approach with appropriate documentation.

Patch-set: 3
Label: Code-Review=+2