Implement Depends-On for github
This adds the capability for one github project pr to depend on another
project's PR, by marking it so in the commit message. This requires
fetching all the commit messages for every commit in a pull request to
see if any of them depend on something.

Various bits of how change objects are created and cached needed to be
fixed to support a change being added while adding another change.

Also put a try around the webhook event handling, so that a traceback,
such as when a dependency loop is detected, can be gracefully logged
instead of the whole thing shutting down.

Reverse searching for Needed By changes is not currently implemented,
due to a limitation in the github search API. More discussion will
happen around the convenience feature of re-enqueuing previous changes.

Change-Id: Ia32e6cc70b163f7e32e74d88f0a1f9b4e3255320
Story: 2000774
Task: 4686