Add ROADMAP.rst file
There is a clear current roadmap in the README, but it's mostly focused
on the "get v3 out the door" stuff. There is a decent backlog of other
things that's useful to be able to point to as "yup, aware of, purposely
deferred" - as well as a set of things that are lower priority right now
but still in-scope for current efforts.

Add all of those to a file called ROADMAP.rst.

While moving stuff over, removed the shim tasks, as I'm pretty sure
we've given up on those, and the vague container related tasks, which
now have more expanded entries.

Change-Id: I07c64a3de14868c1f63a951e6e011d08fa4c8e60