Add ability to auto-generate simple one-line shell playbooks
Instead of requiring a boilerplate playbook for simple one-liner
commands, add ability to just say "run this one-line command" directly
in the zuul.yaml. Although we don't want zuul.yaml to grow playbook-like
logic (that's what playbooks are for) there are plenty of times in
testing/CI where what the user wants is either a really simple shell
command and does not need a playbook to run it.

Don't allow any flexibility. If the user wants to pass arguments or
or do literally ANYTING even mildy complex, they need a playbook.

Change-Id: Ic82c9e67f87f8747129cab1eaae7e379efbbe8c9