Common threading layer
A problem I always have with zuul is that a misconfiguration or failure
will crash a thread. The thread will then exit without logging anything
and zuul will be running but not respond as expected.

Standardize the way we create and load threads. This makes sure that we
always daemonize our threads so that the application can exit cleanly
and will catch and log any failures.

Change-Id: I490e621115b49c194be8558b6d4140f909ffa8dc