Update patch set 3
Patch Set 3: Code-Review+2

Discussed py2 with Jim on irc and we agreed it's fine to do it as a best-effort. The other changes requested will make the program capable of encrypting binary secrets, which seems like a win, except that I'm pretty sure we figured out Ansible fails at this because it has no way to pass binary variables to modules. So, this is fine as-is, though I'd love to see it tested. Did I say test? test. :)

Patch-set: 3
Reviewer: Gerrit User 6488 <6488@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2