Update patch set 1
Patch Set 1:

Wouldn't this be equally possible by creating a base job that is a "shell" job, calling the job and listing an argument that is the script to execute, or the line of shell to execute? Saves us from trying to autogen a playbook for this.

Patch-set: 1
Reviewer: Gerrit User 6589 <6589@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0