Update patch set 1
Patch Set 1: Code-Review-1

If I'm not mistaken the "user" parameter is used by the executor to drop priveleges and change to the specified user if it is started as root.  That's distinct from the "default_username" parameter which is used to log into nodes.

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1