Write secrets into their own file, not into inventory
Publishing the inventory as part of publishing logs would be handy
in debugging that everything is as one expects. However, secrets
currently go into the inventory, which means publishing it is not safe.

Write the secrets to their own file and  only pass it to ansible-playbook
if the playbook is trusted.

Change-Id: If58b89882a817ff219ed5f8faf2bde31cc8e1a6a