Update patch set 1
Patch Set 1: Code-Review+1

It's a little strange that we are relaxing "no local actions" for the stat module. Where would we document such an exception to the rule? What are the scenarios in which this is needed?

Patch-set: 1
Reviewer: Gerrit User 6589 <6589@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1