Update patch set 7
Patch Set 7:

(1 comment)

There's a bunch of things in here that seem like they're good candidates for anchors and references.

There's also a thing we did over in os-api-ref (the "make REST docs for openstack sphinx plugin") for describing nested attributes that might be worth considering (or might be a terrible idea in this content) Namely - keeping a prefix on each structure member, so that someone reading the docs has context clues- but bolding the actual term in question.

For instance, **success** in the pipeline section would be:

pipeline.**success**

And **zuul** in the roles section of job would be:

job.roles[].**zuul""

Patch-set: 7