Update patch set 7
Patch Set 7:

(1 comment)

Agreed -- every keyword should have an anchor.  We may want to make a sphinx role to accomplish that.

I like the prefix idea.  I might be inclined to dawdle a bit on whether the pseudo-syntax of "job.roles[].zuul" is necessary or if "job.roles.zuul" might be just as useful for these purposes, but I think either would be an improvement.

Patch-set: 7