Add html based websocket client for console stream
This adds a static html page to zuul-web that a browser can be pointed
to for streaming log files. To leverage this in the status UI the
scheduler sets the build url to this html page adding build uuid,
logfile and optionally a different url for accessing the
websocket.

Tobias has to run his websocket streamer on a different domain than the
other things, via proxy things - so the url zuul-web is serving for the
static file isn't the same as what it is for the websocket from a
consumer perspective. So introduce a config variable for zuul-web
that allows setting an explicit url for that. If it's not set, use the
relative path from static/stream.html to console-stream.

There should be no issue with relative paths between static/stream.html
and other paths served by zuul-web, so introduce an info endpoint that
stream.html can hit to grab config options, like websocket_url. Have
that return either the configured or the generated url.

Further to not throwing away the finger url retail this as additional
field in status.json. With this a later change to the status ui could
let the user choose between html and finger log streaming.

Co-Authored-By: David Shrewsbury <shrewsbury.dave@gmail.com>
Co-Authored-By: Monty Taylor <mordred@inaugust.com>
Co-Authored-By: Tobias Henkel <tobias.henkel@bmw.de>
Change-Id: I2da7979f448934abe3d41f3a5e50d09004fcccc2