Update patch set 1
Patch Set 1: Code-Review-1

There is a deeper issue at work here...

Jim is working on a fix to how builds are retried. The RETRY_LIMIT issue is related to how we're doing that right now. (which is that we throw retry limit at the start of a new build, rather than at the end of the last failed build) This is also a problem with log uploading - because we aren't running the post playbooks/log uploading for the final failed attempt.

For NOT_REGISTERED - we shouldn't have that any more in v3, right?

Patch-set: 1
Reviewer: Gerrit User 2 <2@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1