Update patch set 1
Patch Set 1:

Do we need to do a format pattern?  Can we just set the hostname in the config file, and since we know the url scheme for the web streamer, construct the url based on that and the configured hostname?

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0