Update patch set 2
Patch Set 2: Code-Review+2

This patch doesn't change anything related to the following, but I thought it worth mentioning: I'd actually like to move away from the is-a thread model to the has-a thread model (I haven't added any new is-a threads in years).  I think the has-a model is a bit more explicit, and we don't have to worry as much about overriding the base thread class.  That would let us drop the implied run/exec stuff.

But, as I said, not strictly related to this patch.  This looks good.

Patch-set: 2