Add job's project as implicit role project
Add the project in which a job is defined as an implicit role project.
This is a convenience for job authors who may want to put roles in
the root of the project (ie, not adjacent to job playbooks, since,
after all, the roles may be useful outside of the job playbooks).  In
that case, they will not need to specify the job's own project in the
roles: section of the job.

Change-Id: Ia382c2da9f7eb7139ceb0b61cb986aace8dc8d8f