Fix error loading projects out of order
When a config altering change lands, we reload the tenant's
configuration using cached data except for the project-branch in
question.  Due to an error, we loaded all of the cached data before
the newly changed data.  Correct this to always load data in the
same sequence regardless of whether it was cached.

Change-Id: If8adc9f3e06474f9bcf8363ae1b64b7f42334a11