Update patch set 1
Patch Set 1: Code-Review-1

I think there was an IRC discussion about this where a discrepancy was noted between the docs and current behavior.  I think we allow both zuul.yaml and .zuul.yaml in all kinds of repos now, but forgot to update the docs.

Since we're changing that section here anyway, and that was the intent, I'd appreciate it if you would go ahead and correct the docs to indicate that all 4 things can be used in all repos.  That should mostly be deleting text -- the only thing we need to mention related to the difference between project types is that we only read the master branch on config repos.

Patch-set: 1
Label: Code-Review=-1