Add spacing lines and formatting to result dicts
Based on a suggestion from pabelanger to try out some of the formatting
ideas from here:

  http://logs.openstack.org/89/481589/1/check/gate-windmill-deploy-fedora-25-nv/493e687/console.html.gz

add blank lines for spacing at the start of tasks and plays, and format
the results dict when we have one. Use yaml for the formatting though,
as we're formatting for for humans to read.

In local testing it seems more readable.

Change-Id: I29945d5451c6ac69c90a0140e1a686ae3afea340